backend/requirements.txt
fastapi==0.117.1
mysql_connector_repackaged==0.3.1
